fix(codegen): fix the union regression in the null walk
The path walk that tells ruby/rust/swift where a null means "explicitly
null" rather than "unset" read only a schema's own `properties`, so a
union arm's key was invisible to it. quicktype MERGES the branches of an
`anyOf` into one class, and it marks a property required only when every
branch requires it — so the inactive arm of a two-branch union rendered
as an unset optional whose null was then sent to the server as an
explicit null, failing validation on a key the caller never set.

The walk now models that merge: it intersects `required` across the
object branches of each combinator, treats a `{"type": "null"}` branch as
contributing no shape, and bounds itself at depth 32 so a
self-referential `$ref` cannot spin. Dart's fail-closed constructor guard
stays — it catches the same class of defect from the other side, at
render time, naming the class it could not reconcile.

The paths now hang off `SdkMethod` where the rest of a method's derived
data lives, instead of a parallel per-model map threaded through every
target's render input; five of the eight targets never read it.

Split the Dart client, which had grown to 1,325 lines, into the four
concerns it had accumulated — shapes, transport, replay, and the client
itself (729) — and split the single conformance file the same way. The
runner now counts its own cases instead of asserting a hand-maintained
total that had already drifted.
