+++ packages/codegen/__tests__/sdk-null-paths.test.ts
@@ -7,7 +7,7 @@
 import { describe, expect, it } from "vitest";
 
 import type { OpenRpcDocument } from "../src/sdk";
-import { modelNullPaths, nullPathsOf } from "../src/sdk/spec";
+import { nullPathsOf, parseSpec } from "../src/sdk/spec";
 
 /** `.nullable()` renders as this, and so does `v.union(v.null(), x)`. */
 const nullable = (inner: Record<string, unknown>): Record<string, unknown> => {
@@ -102,6 +102,74 @@ describe("nullPathsOf", () => {
         expect(paths.nullable).toStrictEqual([["rows", "*", "note"]]);
     });
 
+    it("treats a union's properties the way quicktype MERGES them", () => {
+        expect.assertions(2);
+
+        // The regression this exists to prevent. quicktype folds an `anyOf` into
+        // ONE class whose every property is nullable unless every branch requires
+        // it, so a model built from `v.union(v.object({a}), v.object({b}))` emits
+        // BOTH — one of them null. Reading each branch's own `required` in
+        // isolation would call both required, prune nothing, and send
+        // `{"a":"x","b":null}`, which neither branch of the union accepts.
+        const paths = nullPathsOf({
+            anyOf: [
+                { additionalProperties: false, properties: { a: { type: "string" } }, required: ["a"], type: "object" },
+                { additionalProperties: false, properties: { b: { type: "number" } }, required: ["b"], type: "object" },
+            ],
+        });
+
+        expect(paths.optional).toStrictEqual([["a"], ["b"]]);
+        expect(paths.nullable).toStrictEqual([]);
+    });
+
+    it("keeps a property required when EVERY branch requires it", () => {
+        expect.assertions(2);
+
+        const paths = nullPathsOf({
+            anyOf: [
+                { properties: { kind: { type: "string" }, only: { type: "string" } }, required: ["kind", "only"], type: "object" },
+                { properties: { kind: { type: "string" } }, required: ["kind"], type: "object" },
+            ],
+        });
+
+        // `kind` is in both; `only` is missing from the second, which is exactly
+        // what makes it optional in the merged class.
+        expect(paths.optional).toStrictEqual([["only"]]);
+        expect(paths.nullable).toStrictEqual([]);
+    });
+
+    it("does not let a nullable wrapper make an object's properties optional", () => {
+        expect.assertions(2);
+
+        // `v.object({…}).nullable()` is an `anyOf` too, but its `{type:"null"}`
+        // branch is not an object and contributes no shape — so intersecting must
+        // not sweep the real branch's requirements away with it.
+        const paths = nullPathsOf({
+            properties: { profile: nullable({ properties: { bio: nullable({ type: "string" }) }, required: ["bio"], type: "object" }) },
+            required: ["profile"],
+            type: "object",
+        });
+
+        expect(paths.optional).toStrictEqual([]);
+        expect(paths.nullable).toStrictEqual([["profile"], ["profile", "bio"]]);
+    });
+
+    it("treats allOf as a conjunction, not a choice", () => {
+        expect.assertions(2);
+
+        // Every `allOf` branch holds at once, so a property required by any of
+        // them is required — the opposite of the `anyOf` rule above.
+        const paths = nullPathsOf({
+            allOf: [
+                { properties: { a: { type: "string" } }, required: ["a"], type: "object" },
+                { properties: { b: { type: "string" } }, required: ["b"], type: "object" },
+            ],
+        });
+
+        expect(paths.optional).toStrictEqual([]);
+        expect(paths.nullable).toStrictEqual([]);
+    });
+
     it("does not walk `additionalProperties: false` as if it were a schema", () => {
         expect.assertions(2);
 
@@ -160,8 +228,8 @@ describe("nullPathsOf", () => {
     });
 });
 
-describe("modelNullPaths", () => {
-    it("keys the paths by the same model name the backend renders", () => {
+describe("argsNullPaths on a parsed method", () => {
+    it("carries the paths for the method's own arguments", () => {
         expect.assertions(2);
 
         const document: OpenRpcDocument = {
@@ -184,11 +252,11 @@ describe("modelNullPaths", () => {
             ],
         };
 
-        const paths = modelNullPaths(document);
+        // On the method, so a target reads them where it already has the method
+        // — no lookup, and no second naming rule to keep in step.
+        const method = parseSpec(document)[0]?.methods[0];
 
-        // The name `parseMethod` predicts, so a target can look it up by a
-        // method's `argsType` with no second naming rule.
-        expect(paths["RecordsPutArgs"]?.optional).toStrictEqual([["limit"]]);
-        expect(paths["RecordsPutArgs"]?.nullable).toStrictEqual([["nickname"]]);
+        expect(method?.argsNullPaths.optional).toStrictEqual([["limit"]]);
+        expect(method?.argsNullPaths.nullable).toStrictEqual([["nickname"]]);
     });
 });
