+++ sdks/dart/lib/lunora.dart
@@ -1,11 +1,15 @@
 /// The hand-written Lunora transport for Dart and Flutter.
 ///
-/// Three pieces, and nothing else: the tagged value codec (`src/wire.dart`), the
-/// stable subscription key (`src/key.dart`), and the RPC + WebSocket client
-/// (`src/client.dart`). It imports `dart:convert`, `dart:typed_data` and
-/// `dart:async` and nothing more, so it runs unchanged on every Flutter target
-/// — iOS, Android, web, macOS, Windows, Linux — with no FFI and no conditional
-/// import.
+/// The wire codec (`src/wire.dart`), the stable subscription key
+/// (`src/key.dart`), the RPC + WebSocket client (`src/client.dart`), and the two
+/// pieces that make it usable offline: the optimistic-update engine
+/// (`src/optimistic.dart`) and the mutation queue (`src/offline_queue.dart`),
+/// with the shape/poke protocol in `src/shapes.dart` and their coded errors in
+/// `src/errors.dart`.
+///
+/// It imports `dart:convert`, `dart:typed_data` and `dart:async` and nothing
+/// more, so it runs unchanged on every Flutter target — iOS, Android, web,
+/// macOS, Windows, Linux — with no FFI and no conditional import.
 ///
 /// `lunora sdk generate --lang dart` copies this `lib/` into the output beside
 /// the generated surface, so nothing here may refer to its own package by name:
@@ -18,4 +22,7 @@ export 'src/errors.dart';
 export 'src/key.dart';
 export 'src/offline_queue.dart';
 export 'src/optimistic.dart';
+export 'src/replay.dart';
+export 'src/shapes.dart';
+export 'src/transport.dart';
 export 'src/wire.dart';
